pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "road-lane-segmentation"
version = "0.1.0"
description = "Awesome `road-lane-segmentation` project!"
readme = "README.md"
authors = ["Awesome user <awesome-user@example.com>"]
repository = ""
homepage = ""

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "~3.9.0"
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^10.7.0"
python-dotenv = "^0.21.1"
torch = "^1.13.1"
pandas = "^1.5.3"
numpy = "^1.24.2"
opencv-python-headless = "^4.7.0.72"
albumentations = "^1.3.0"
openpyxl = "^3.1.1"
hydra-core = "^1.3.2"
checksumdir = "^1.2.0"
catalyst = ">=21.10,<22.0"
torchvision = "^0.14.1"
segmentation-models-pytorch = "^0.3.2"
pyyaml = "^6.0"
tensorboardx = "^2.6"
imageio = "^2.26.0"
imageio-ffmpeg = "^0.4.8"

[tool.poetry.group.dev.dependencies]
black = {version = "^21.7", allow-prereleases = true}
isort = {extras = ["colors"], version = "^5.10.1"}
pre-commit = "^2.14.0"
nbdime = "^3.1.0"
envyaml = "^1.9.210927"
ipykernel = "^6.4.2"
jupyter = "^1.0.0"
jinja2 = "3.0.3"
matplotlib = "^3.7.0"
tqdm = "^4.64.1"
clearml = "^1.9.1"
jupyter-contrib-nbextensions = "^0.7.0"
jupyter-client = "^8.0.3"
jupyterlab = "^3.6.1"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 100
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 100

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]
